bzero changed to allow unsigned long size

+++ get_next_line.h
@@ -6,7 +6,7 @@
 /*   By: albgarci <albgarci@student.42madrid.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/03 20:29:39 by albgarci          #+#    #+#             */
-/*   Updated: 2021/10/03 20:29:57 by albgarci         ###   ########.fr       */
+/*   Updated: 2021/10/13 12:34:08 by albgarci         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@ size_t	ft_strlen(const char *s);
 void	*ft_memmove(void *dst, const void *src, size_t len);
 int		ft_c(const char *s, int c);
 char	*ft_substr(char const *s, unsigned int start, size_t len);
-void	ft_bzero(void *s, size_t n);
+void	ft_bzero(void *s, unsigned long n);
 char	*ft_first_read(int fd, char *lft);
 int		ft_read_loop(int fd, char **lft);
 char	*ft_create_line(char **lft);
